+++ addons/web_editor/static/src/js/editor/odoo-editor.js
@@ -1,1 +1,1 @@
-/home/goaman/projects/external/odoo/odoo-editor/build/odoo-editor-bundle.js
+/home/odoo/Development/odoo-work/odoo-editor/build/odoo-editor-bundle.js
