+++ addons/website/views/assets.xml
@@ -154,8 +154,6 @@
     <link rel="stylesheet" type="text/scss" href="/website/static/src/scss/website.editor.ui.scss"/>
     <link rel="stylesheet" type="text/scss" href="/website/static/src/scss/website.theme_install.scss"/>
 
-    <script type="text/javascript" src="/website/static/src/js/editor/editor_menu_translate.js"/>
-
     <script type="text/javascript" src="/website/static/src/js/menu/content.js"/>
     <script type="text/javascript" src="/website/static/src/js/menu/customize.js"/>
     <script type="text/javascript" src="/website/static/src/js/menu/debug_manager.js"/>
